Code cleanup: deduplicate crapto1 library - merge the existing 4 crapto1 sources into common/crapto1
