+++ armsrc/mifaresniff.h
@@ -11,29 +11,22 @@
 #ifndef __MIFARESNIFF_H
 #define __MIFARESNIFF_H
 
-#include "proxmark3.h"
-#include "apps.h"
+#include <stdint.h>
+#include <stdbool.h>
 #include "util.h"
-#include "string.h"
-
-#include "iso14443crc.h"
-#include "iso14443a.h"
-#include "crapto1.h"
-#include "mifareutil.h"
-#include "common.h"
 
 #define SNF_INIT				0
-#define SNF_NO_FIELD		1
-#define SNF_WUPREQ			2
+#define SNF_NO_FIELD			1
+#define SNF_WUPREQ				2
 #define SNF_ATQA				3
-#define SNF_ANTICOL1		4
+#define SNF_ANTICOL1			4
 #define SNF_UID1				5
-#define SNF_ANTICOL2		6
+#define SNF_ANTICOL2			6
 #define SNF_UID2				7
 #define SNF_SAK					8
-#define SNF_CARD_IDLE		9
-#define SNF_CARD_CMD		10
-#define SNF_CARD_RESP		11
+#define SNF_CARD_IDLE			9
+#define SNF_CARD_CMD			10
+#define SNF_CARD_RESP			11
 
 #define SNF_UID_4				0
 #define SNF_UID_7				0
