+++ common/crapto1/crypto1.c
@@ -23,21 +23,38 @@
 #define SWAPENDIAN(x)\
 	(x = (x >> 8 & 0xff00ff) | (x & 0xff00ff) << 8, x = x >> 16 | x << 16)
 
+#if defined(__arm__)
+void crypto1_create(struct Crypto1State *s, uint64_t key)
+{
+#else
 struct Crypto1State * crypto1_create(uint64_t key)
 {
 	struct Crypto1State *s = malloc(sizeof(*s));
+#endif
 	int i;
 
 	for(i = 47;s && i > 0; i -= 2) {
 		s->odd  = s->odd  << 1 | BIT(key, (i - 1) ^ 7);
 		s->even = s->even << 1 | BIT(key, i ^ 7);
 	}
+#if defined(__arm__)	
+	return;
+#else
 	return s;
+#endif
 }
+#if defined(__arm__)
+void crypto1_destroy(struct Crypto1State *state)
+{
+	state->odd = 0;
+	state->even = 0;
+}
+#else
 void crypto1_destroy(struct Crypto1State *state)
 {
 	free(state);
 }
+#endif
 void crypto1_get_lfsr(struct Crypto1State *state, uint64_t *lfsr)
 {
 	int i;
